+++ sdk/storage/azure-storage-blobs/CMakeLists.txt
@@ -41,7 +41,7 @@ endif()
 set(
   AZURE_STORAGE_BLOBS_HEADER
     inc/azure/storage/blobs/append_blob_client.hpp
-    inc/azure/storage/blobs/blob_batch_client.hpp
+    inc/azure/storage/blobs/blob_batch.hpp
     inc/azure/storage/blobs/blob_client.hpp
     inc/azure/storage/blobs/blob_container_client.hpp
     inc/azure/storage/blobs/blob_lease_client.hpp
@@ -63,7 +63,7 @@ set(
 set(
   AZURE_STORAGE_BLOBS_SOURCE
     src/append_blob_client.cpp
-    src/blob_batch_client.cpp
+    src/blob_batch.cpp
     src/blob_client.cpp
     src/blob_container_client.cpp
     src/blob_lease_client.cpp
