+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_client.hpp
@@ -431,6 +431,6 @@ namespace Azure { namespace Storage { namespace Blobs {
     friend class Files::DataLake::DataLakeDirectoryClient;
     friend class Files::DataLake::DataLakeFileClient;
     friend class BlobLeaseClient;
-    friend class BlobBatchClient;
+    friend class BlobBatch;
   };
 }}} // namespace Azure::Storage::Blobs
